# src/main/java/com/inditex/priceservice/infrastructure/in/rest/PriceResponse.java
package com.inditex.priceservice.infrastructure.in.rest;

import java.math.BigDecimal;
import java.time.LocalDateTime;

public record PriceResponse(
        Long productId,
        Integer brandId,
        Integer priceList,
        LocalDateTime startDate,
        LocalDateTime endDate,
        BigDecimal price,
        String currency
) {}