# src/main/resources/openapi/price-api.yaml
openapi: 3.0.3
info:
  title: Price Service API
  version: 1.0.0

paths:
  /prices:
    get:
      tags:
        - Prices
      summary: Get applicable price
      operationId: getApplicablePrice
      parameters:
        - name: applicationDate
          in: query
          required: true
          description: Date and time of application
          schema:
            type: string
            format: date-time
            example: "2026-06-05T10:00:00"
        - name: productId
          in: query
          required: true
          description: Product id
          schema:
            type: integer
            format: int64
            example: 12345
        - name: brandId
          in: query
          required: true
          description: Brand id
          schema:
            type: integer
            format: int32
            example: 1
      responses:
        "200":
          description: Applicable price found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResponse'
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No applicable price found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    PriceResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: Product id
          example: 12345
        brandId:
          type: integer
          format: int32
          description: Brand id
          example: 1
        priceList:
          type: integer
          format: int32
          description: Price list id
          example: 1
        startDate:
          type: string
          format: date-time
          description: Start date
          example: "2020-06-14T00:00:00"
        endDate:
          type: string
          format: date-time
          description: End date
          example: "2020-12-31T23:59:59"
        price:
          type: number
          format: double
          description: Final price
          example: 35.50
        currency:
          type: string
          description: Currency code
          example: EUR

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        message:
          type: string
          example: "No applicable price found for the given parameters"